requirements.txt
# requirements.txt
requests
Flask
hubspot-api-client
py3-validate-email
